Replace TOC scroll logic with pathname-based nav
